+++ ljh/.gitignore
@@ -1,0 +1,26 @@
+
+/bin/
+
+JAVA/.metadata/
+
+JAVA/bin/
+
+JAVA/.settings/
+
+JAVA/.classpath
+
+JAVA/.project
+
+JavaScript/coverage/
+
+JavaScript/node_modules/
+
+JavaScript/babel.config.json
+
+JavaScript/jest.config.json
+
+JavaScript/package.json
+
+JavaScript/package-lock.json
+
+/.metadata/
