+++ ljh/java/src/level1/HateSameNumber.java
@@ -1,4 +1,4 @@
-package level1.java;
+package level1;
 
 /*
  * 프로그래머스 - 같은 숫자는 싫어
