+++ ljh/java/src/level1/SumBetweenTwoNumber.java
@@ -1,4 +1,4 @@
-package level1.java;
+package level1;
 
 /*
  * 프로그래머스 - 두 정수 사이의 합
