+++ ljh/javascript/level1/DivideArrayNumber.js
