+++ ljh/javascript/level1/HateSameNumber.js
