--- ljh/src/level1/js/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<script src="SumNumber.js"></script>
-<title></title>
-</head>
-<body>
-
-</body>
-</html>
